refactor: reunion with parents
